requirements.txt
ebooklib
bs4
openai
python-dotenv
moviepy
requests
streamlit
jupyter
beautifulsoup4
pillow